DW/app/requirements.txt
Django>=2.1,<2.2
wagtail>=2.4,<2.5
gunicorn>=19.7,<19.8
mysqlclient>=1.3.12,<1.4
# wagtailmedia>=0.2.0,<0.3
git+https://github.com/evanwinter/wagtailmedia.git@master
django-storages>=1.6.6,<2.0
boto3>=1.7.70,<2.0
instaloader>=4.1.1,<5.0
django-favicon-plus>=0.0.8,<0.1
wagtail-react-streamfield>=1.1,<1.2
wagtailfontawesome>=1.1.3,<1.2
